+++ gflow/cli/__init__.py
@@ -11,29 +11,3 @@
 #
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-from dataclasses import asdict, dataclass
-
-import requests
-import typer
-
-from gflow_cli.constants import DATASETS_URL
-
-app = typer.Typer(help="Manage your datasets with `gflow-cli datasets` command.")
-
-AVAILABLE_STORAGE = ("LOCAL", "REMOTE")
-AVAILABLE_TASKS = ("IMAGE-CLASSIFICATION",)
-AVAILABLE_DATASET_TYPES = ("FROM-FOLDER", "FROM-CSV")
-
-
-@dataclass
-class DatasetRequest:
-    path: str
-    task: str
-    dataset_type: str
-    remote: bool
-
-
-@app.command(name="available-tasks")
-def get_available_tasks() -> None:
-    typer.echo(AVAILABLE_TASKS)
