+++ sonar-project.properties
@@ -6,7 +6,7 @@ sonar.organization=gflow
 #sonar.projectVersion=1.0
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=gflow_cli
+sonar.sources=gflow
 
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
