+++ tests/test_projects.py
@@ -17,16 +17,16 @@ from unittest.mock import MagicMock, patch
 import typer
 from typer.testing import CliRunner
 
-from gflow_cli.projects import add_project
-from gflow_cli.utility import init_config
+from gflow.cli.projects import add_project
+from gflow.utility import init_config
 
 runner = CliRunner()
 
 init_config("fake", "fake")
 
 
-@patch("gflow_cli.projects.requests.post")
-@patch("gflow_cli.projects.read_config")
+@patch("gflow.projects.requests.post")
+@patch("gflow.projects.read_config")
 def test_add_project(mock_read_config: MagicMock, mock_post: MagicMock):
     mock_response = mock_post.return_value = MagicMock()
     mock_response.return_value = False
